backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
google-generativeai==0.8.4
yfinance==0.2.51
pandas==2.2.3
numpy==2.2.1
aiosqlite==0.20.0
websockets==14.1
